STABLE14-resync-20041018
cheat. pull in all the 1.3.72->head changes in one delta

+++ src/util/snprintf.c
@@ -436,9 +436,11 @@ afs_vsnprintf(char *p, size_t avail, const char *fmt, va_list ap)
 		} else {
 		    x = "%u.%u.%u.%u";
 		}
-		sprintf(xbuf, x, (UVAL & 0xff000000) >> 24,
-			(UVAL & 0x00ff0000) >> 16, (UVAL & 0x0000ff00) >> 8,
-			(UVAL & 0x000000ff));
+		/* typecast to whatever '%u' is! */
+		sprintf(xbuf, x, (unsigned int)((UVAL & 0xff000000) >> 24),
+			(unsigned int)((UVAL & 0x00ff0000) >> 16), 
+			(unsigned int)((UVAL & 0x0000ff00) >> 8),
+			(unsigned int)(UVAL & 0x000000ff));
 		x = xbuf;
 		len = strlen(xbuf);
 	    }
